backend/requirements.txt
# ── DJANGO ────────────────────────────────────────────────
django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
python-decouple==3.8

# ── ASYNC / CELERY ────────────────────────────────────────
celery==5.3.6
redis==5.0.4
django-celery-results==2.5.1

# ── AUTH KEYCLOAK ─────────────────────────────────────────
python-keycloak==3.9.1
PyJWT==2.8.0

# ── IA — LANGCHAIN / LANGGRAPH ────────────────────────────
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langgraph==0.2.14
langsmith==0.1.98

# ── LLM — OLLAMA ──────────────────────────────────────────
langchain-ollama==0.1.3

# ── DATA ──────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2

# ── RAG / VECTORSTORE ─────────────────────────────────────
chromadb==0.5.0
sentence-transformers==3.0.1

# ── PDF ───────────────────────────────────────────────────
pdfplumber==0.11.1
pytesseract==0.3.10
Pillow==10.3.0

# ── TESTS ─────────────────────────────────────────────────
pytest==8.2.0
pytest-django==4.8.0
pytest-asyncio==0.23.6
factory-boy==3.3.0